refactor(mobile): mobile ディレクトリのコードを綺麗にする
