+++ mobile/assets/images/onboarding/overview-screen-background.png
