+++ mobile/src/app/AuthGate.tsx
