+++ mobile/src/features/history/hooks/useJudgmentDetail.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 
 import { getJudgmentDetail } from '@/features/history/api/judgmentApi';
-import type { Judgment } from '@/features/session/types';
 import type { ApiError } from '@/shared/lib/api';
+import type { Judgment } from '@/shared/types/session';
 
 export function useJudgmentDetail(judgmentId: string) {
   return useQuery<Judgment, ApiError>({
