+++ mobile/src/features/history/hooks/useJudgments.ts
@@ -4,9 +4,9 @@
 import { useQuery } from '@tanstack/react-query';
 
 import { listJudgments } from '@/features/history/api/judgmentApi';
-import type { Judgment } from '@/features/session/types';
 import type { ApiError } from '@/shared/lib/api';
 import type { Paginated } from '@/shared/types/api';
+import type { Judgment } from '@/shared/types/session';
 
 export function useJudgments() {
   return useQuery<Paginated<Judgment>, ApiError>({
