+++ mobile/src/features/session/hooks/useNextCycleRotation.ts
@@ -1,0 +1,213 @@
+import { useCallback, useEffect, useMemo, useRef } from 'react';
+import {
+  Animated,
+  Easing,
+  type GestureResponderEvent,
+  type LayoutChangeEvent,
+  PanResponder,
+} from 'react-native';
+
+const NEXT_CYCLE_IDLE_ROTATION_DEGREES = 5;
+const NEXT_CYCLE_ROTATE_THRESHOLD_DEGREES = 360;
+const NEXT_CYCLE_MAX_DRAG_ROTATION_DEGREES = 1080;
+const NEXT_CYCLE_ROTATION_SENSITIVITY = 1.25;
+const NEXT_CYCLE_PATH_ROTATION_DEGREES_PER_PIXEL = 1.15;
+const NEXT_CYCLE_MIN_ROTATION_RADIUS = 48;
+const NEXT_CYCLE_ROTATION_AREA_FALLBACK = { width: 320, height: 430 };
+
+function clampRotation(rotation: number) {
+  return Math.max(
+    -NEXT_CYCLE_MAX_DRAG_ROTATION_DEGREES,
+    Math.min(NEXT_CYCLE_MAX_DRAG_ROTATION_DEGREES, rotation),
+  );
+}
+
+function normalizeRotationDelta(delta: number) {
+  if (delta > 180) return delta - 360;
+  if (delta < -180) return delta + 360;
+  return delta;
+}
+
+function getGestureAngle(
+  event: GestureResponderEvent,
+  areaSize: { width: number; height: number },
+) {
+  const { locationX, locationY } = event.nativeEvent;
+  if (!Number.isFinite(locationX) || !Number.isFinite(locationY)) return null;
+
+  const dx = locationX - areaSize.width / 2;
+  const dy = locationY - areaSize.height / 2;
+  if (Math.hypot(dx, dy) < NEXT_CYCLE_MIN_ROTATION_RADIUS) return null;
+
+  return (Math.atan2(dy, dx) * 180) / Math.PI;
+}
+
+function getGesturePoint(event: GestureResponderEvent) {
+  const { locationX, locationY, pageX, pageY } = event.nativeEvent;
+  if (Number.isFinite(pageX) && Number.isFinite(pageY)) {
+    return { x: pageX, y: pageY };
+  }
+  if (Number.isFinite(locationX) && Number.isFinite(locationY)) {
+    return { x: locationX, y: locationY };
+  }
+
+  return null;
+}
+
+export function useNextCycleRotation({
+  isStarting,
+  onStart,
+}: {
+  isStarting: boolean;
+  onStart: () => void;
+}) {
+  const hourglassRotation = useRef(new Animated.Value(-NEXT_CYCLE_IDLE_ROTATION_DEGREES)).current;
+  const idleAnimation = useRef<Animated.CompositeAnimation | null>(null);
+  const rotationAreaSize = useRef(NEXT_CYCLE_ROTATION_AREA_FALLBACK);
+  const lastTouchAngle = useRef<number | null>(null);
+  const lastTouchPoint = useRef<{ x: number; y: number } | null>(null);
+  const draggedRotation = useRef(0);
+  const pathRotation = useRef(0);
+  const hasCompletedRotationGesture = useRef(false);
+  const hasTriggeredRotationStart = useRef(false);
+
+  const startIdleAnimation = useCallback(() => {
+    idleAnimation.current?.stop();
+    hourglassRotation.setValue(-NEXT_CYCLE_IDLE_ROTATION_DEGREES);
+    const animation = Animated.loop(
+      Animated.sequence([
+        Animated.timing(hourglassRotation, {
+          toValue: NEXT_CYCLE_IDLE_ROTATION_DEGREES,
+          duration: 1400,
+          easing: Easing.inOut(Easing.ease),
+          useNativeDriver: true,
+        }),
+        Animated.timing(hourglassRotation, {
+          toValue: -NEXT_CYCLE_IDLE_ROTATION_DEGREES,
+          duration: 1400,
+          easing: Easing.inOut(Easing.ease),
+          useNativeDriver: true,
+        }),
+      ]),
+    );
+
+    idleAnimation.current = animation;
+    animation.start();
+  }, [hourglassRotation]);
+
+  useEffect(() => {
+    startIdleAnimation();
+    return () => {
+      idleAnimation.current?.stop();
+    };
+  }, [startIdleAnimation]);
+
+  const hourglassRotationStyle = hourglassRotation.interpolate({
+    inputRange: [-NEXT_CYCLE_MAX_DRAG_ROTATION_DEGREES, NEXT_CYCLE_MAX_DRAG_ROTATION_DEGREES],
+    outputRange: [
+      `${-NEXT_CYCLE_MAX_DRAG_ROTATION_DEGREES}deg`,
+      `${NEXT_CYCLE_MAX_DRAG_ROTATION_DEGREES}deg`,
+    ],
+  });
+
+  const triggerNextCycleByRotation = useCallback(() => {
+    if (hasTriggeredRotationStart.current || isStarting) return;
+
+    hasTriggeredRotationStart.current = true;
+    onStart();
+  }, [isStarting, onStart]);
+
+  const handleRotationAreaLayout = useCallback((event: LayoutChangeEvent) => {
+    const { width, height } = event.nativeEvent.layout;
+    if (width <= 0 || height <= 0) return;
+
+    rotationAreaSize.current = { width, height };
+  }, []);
+
+  const rotationResponder = useMemo(
+    () =>
+      PanResponder.create({
+        onStartShouldSetPanResponderCapture: () => !isStarting,
+        onStartShouldSetPanResponder: () => !isStarting,
+        onMoveShouldSetPanResponderCapture: () => !isStarting,
+        onMoveShouldSetPanResponder: () => !isStarting,
+        onPanResponderGrant: (event) => {
+          idleAnimation.current?.stop();
+          lastTouchAngle.current = getGestureAngle(event, rotationAreaSize.current);
+          lastTouchPoint.current = getGesturePoint(event);
+          draggedRotation.current = 0;
+          pathRotation.current = 0;
+          hasCompletedRotationGesture.current = false;
+          hourglassRotation.setValue(0);
+          hasTriggeredRotationStart.current = false;
+        },
+        onPanResponderMove: (event) => {
+          if (isStarting || hasTriggeredRotationStart.current) return;
+
+          const currentPoint = getGesturePoint(event);
+          if (currentPoint !== null) {
+            if (lastTouchPoint.current !== null) {
+              const dx = currentPoint.x - lastTouchPoint.current.x;
+              const dy = currentPoint.y - lastTouchPoint.current.y;
+              pathRotation.current +=
+                Math.hypot(dx, dy) * NEXT_CYCLE_PATH_ROTATION_DEGREES_PER_PIXEL;
+            }
+            lastTouchPoint.current = currentPoint;
+          }
+
+          const currentAngle = getGestureAngle(event, rotationAreaSize.current);
+          if (currentAngle !== null && lastTouchAngle.current === null) {
+            lastTouchAngle.current = currentAngle;
+            return;
+          }
+
+          if (currentAngle !== null && lastTouchAngle.current !== null) {
+            const rotationDelta =
+              normalizeRotationDelta(lastTouchAngle.current - currentAngle) *
+              NEXT_CYCLE_ROTATION_SENSITIVITY;
+            const nextRotation = clampRotation(draggedRotation.current + rotationDelta);
+            draggedRotation.current = nextRotation;
+            hourglassRotation.setValue(nextRotation);
+            lastTouchAngle.current = currentAngle;
+          }
+
+          hasCompletedRotationGesture.current =
+            Math.max(Math.abs(draggedRotation.current), pathRotation.current) >=
+            NEXT_CYCLE_ROTATE_THRESHOLD_DEGREES;
+        },
+        onPanResponderRelease: () => {
+          const shouldStart = hasCompletedRotationGesture.current;
+          lastTouchAngle.current = null;
+          lastTouchPoint.current = null;
+          draggedRotation.current = 0;
+          pathRotation.current = 0;
+          hasCompletedRotationGesture.current = false;
+
+          if (shouldStart) {
+            triggerNextCycleByRotation();
+          } else {
+            hourglassRotation.setValue(0);
+            startIdleAnimation();
+          }
+        },
+        onPanResponderTerminate: () => {
+          lastTouchAngle.current = null;
+          lastTouchPoint.current = null;
+          draggedRotation.current = 0;
+          pathRotation.current = 0;
+          hasCompletedRotationGesture.current = false;
+          if (!hasTriggeredRotationStart.current) {
+            hourglassRotation.setValue(0);
+            startIdleAnimation();
+          }
+        },
+      }),
+    [hourglassRotation, isStarting, startIdleAnimation, triggerNextCycleByRotation],
+  );
+
+  return {
+    handleRotationAreaLayout,
+    hourglassRotationStyle,
+    rotationResponder,
+  };
+}
