+++ mobile/src/features/stats/lib/reportPresentation.ts
@@ -1,0 +1,8 @@
+import { isApiError } from '@/shared/lib/api';
+
+export function getReportErrorMessage(error: unknown): string {
+  if (isApiError(error)) {
+    return error.problem?.detail ?? error.problem?.title ?? 'レポートの取得に失敗しました。';
+  }
+  return 'レポートの取得に失敗しました。';
+}
