+++ mobile/src/shared/lib/colors.ts
@@ -1,0 +1,12 @@
+export const APP_COLORS = {
+  input: '#4B5CFF',
+  output: '#EC4899',
+  stats: '#5367FF',
+  textPrimary: '#2F2F2F',
+  textMuted: '#777777',
+  textInactive: '#9CA3AF',
+  dotInactive: '#D9D9D9',
+  border: '#E5E7EB',
+  error: '#D92D20',
+  white: '#FFFFFF',
+} as const;
